refactor(explore): split explore.ts into explore/ module directory
Split the monolithic src/explore.ts (572 lines, 24KB) into 7 focused modules:

- explore/types.ts     — shared interfaces and type exports
- explore/site.ts      — site name detection and slugify
- explore/network.ts   — network parsing, URL pattern, auth detection
- explore/analysis.ts  — endpoint analysis, response body analysis, scoring
- explore/capabilities.ts — capability inference from analyzed endpoints
- explore/artifacts.ts — write explore artifacts to disk
- explore/index.ts     — main exploreUrl orchestrator + re-exports

Added 50 unit tests covering all extracted pure functions:
- detectSiteName, slugify
- parseNetworkRequests, urlToPattern, detectAuthIndicators
- flattenFields, analyzeResponseBody, scoreEndpoint, analyzeEndpoints
- inferCapabilityName, inferStrategy, inferCapabilitiesFromEndpoints

All 273 tests pass (223 existing + 50 new). Zero TypeScript errors.
